requirements.txt
beautifulsoup4==4.10.0
requests==2.26.0
lxml==4.6.4
dnspython==2.1.0
rich==10.15.2
